pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codegen-ai"
version = "1.0.0"
description = "AI-powered code generation service"
requires-python = ">=3.8"
dependencies = [
    "google-generativeai==0.3.2",
    "python-dotenv==1.0.0",
    "pathlib==1.0.1",
    "requests==2.31.0",
]

[tool.setuptools]
package-dir = {"" = "server"}
packages = ["services"] 